인덱스 관련 포맷 변경

+++ DirectX11/ACamera.cpp
@@ -105,7 +105,7 @@ ACamera::ACamera(
 		{0.f, 0.f}, {1.f, 0.f}, {0.f, 1.f},
 		{0.f, 1.f}, {1.f, 0.f}, {1.f, 1.f},
 	};
-	vector<uint8_t> ResolverIndex{ 0, 1, 2, 3, 4, 5 };
+	vector<uint16_t> ResolverIndex{ 0, 1, 2, 3, 4, 5 };
 
 	//vector<XMFLOAT3> ResolverPositions{ {-1.f, 1.f, 0.f}, {3.f, 1.f, 0.f}, {-1.f, -3.f, 0.f} };
 	//vector<XMFLOAT2> ResolverUVBuffers{ {0.f, 0.f}, {2.f, 0.f}, {0.f, 2.f} };
